+++ packages/react-api/src/api/SQL.js
@@ -2,9 +2,10 @@ import { encodeParameter, getRequest, postRequest } from '@carto/react-core';
 import { REQUEST_GET_MAX_URL_LENGTH } from '@carto/react-core';
 import { API_VERSIONS } from '@deck.gl/carto';
 
-import { dealWithApiError, generateApiUrl } from './common';
+import { dealWithApiError } from './common';
 
 const CLIENT = 'carto-react';
+const DEFAULT_USER_COMPONENT_IN_URL = '{user}';
 
 /**
  * Executes a SQL query
@@ -99,4 +100,48 @@ function createRequest({ credentials, connection, query, opts = {} }) {
     parameters: urlParamsForPost
   });
   return postRequest(postUrl, rawParams, requestOpts);
+}
+
+/**
+ * Generate a valid API url for a request
+ */
+export function generateApiUrl({ credentials, connection, parameters }) {
+  const { apiVersion = API_VERSIONS.V2, apiBaseUrl } = credentials;
+
+  let url;
+  switch (apiVersion) {
+    case API_VERSIONS.V1:
+    case API_VERSIONS.V2:
+      url = `${sqlApiUrl(credentials)}api/v2/sql`;
+      break;
+
+    case API_VERSIONS.V3:
+      url = `${apiBaseUrl}/v3/sql/${connection}/query`;
+      break;
+
+    default:
+      throw new Error(`Unknown apiVersion ${apiVersion}`);
+  }
+
+  if (!parameters) {
+    return url;
+  }
+
+  return `${url}?${parameters.join('&')}`;
+}
+
+/**
+ * Prepare a url valid for the specified user, from the serverUrlTemplate
+ */
+function sqlApiUrl(credentials) {
+  let url = credentials.serverUrlTemplate.replace(
+    DEFAULT_USER_COMPONENT_IN_URL,
+    credentials.username
+  );
+
+  if (!url.endsWith('/')) {
+    url += '/';
+  }
+
+  return url;
 }