+++ packages/react-widgets/__tests__/models/FormulaModel.test.js
@@ -1,7 +1,6 @@
 import { getFormula } from '../../src/models/FormulaModel';
 import { AggregationTypes } from '@carto/react-core';
 import { Methods, executeTask } from '@carto/react-workers';
-import { sources } from 'webpack';
 import { executeSQL } from '@carto/react-api';
 
 const RESULT = 3.14;
@@ -15,7 +14,7 @@ jest.mock('@carto/react-api', () => ({
 jest.mock('@carto/react-workers', () => ({
   executeTask: jest
     .fn()
-    .mockImplementation(() => new Promise((resolve) => resolve(RESULT))),
+    .mockImplementation(() => new Promise((resolve) => resolve({ value: RESULT }))),
   Methods: {
     FEATURES_FORMULA: 'featuresFormula'
   }
@@ -39,7 +38,7 @@ describe('getFormula', () => {
 
       const data = await getFormula(props);
 
-      expect(data).toBe(RESULT);
+      expect(data).toStrictEqual({ value: RESULT });
 
       expect(executeTask).toHaveBeenCalledWith(
         props.source.id,
@@ -75,7 +74,7 @@ describe('getFormula', () => {
 
       const data = await getFormula(props);
 
-      expect(data).toBe(RESULT);
+      expect(data).toStrictEqual({ value: RESULT });
 
       expect(executeSQL).toHaveBeenCalledWith({
         credentials: props.source.credentials,